Added basic install script
The script install_files.sh has been added to install Cask, configuration.org, and tmux.conf files as symbolic links.

Emacs currently picks up on the fact that configuration.org is a symbolic link to a git controlled directory and prompts me on startup if this is okay. I will need to find a workaround to surpress this message as it should
generally be treated as okay. The script also has no intellegence so will fail if presented with preexisting files.