Massive reorganisation
This is now using gnu stow and ansible to control things.

The Ansible playbook installs everything. Stow sets up the symbolic links between my home directory and the dotfiles.

A future issue to cope with is that Wayland will be a thing soon and will break i3 and polybar, etc. :(