Various changes - HEADLINE GTD EMACS
Quite a lot of changes in this, but the major change is that I've
added a GTD workflow to org-mode. Let's see if this makes me happier.

I've also added the inconsolata font to emacs, and ubuntu
installation. And the silversearcher-ag package. (Something I need to
remember to use more frequently)

i3 now only starts emacs as everything is optional depending on the
computer I'm using.